+++ generate/genfile/genfile_test.go
@@ -24,8 +24,6 @@ import (
 	"github.com/mineiros-io/terramate/generate/genfile"
 	"github.com/mineiros-io/terramate/hcl"
 	"github.com/mineiros-io/terramate/hcl/info"
-	"github.com/mineiros-io/terramate/lets"
-	maptest "github.com/mineiros-io/terramate/mapexpr/test"
 	"github.com/mineiros-io/terramate/project"
 	"github.com/mineiros-io/terramate/stack"
 	"github.com/mineiros-io/terramate/test"
@@ -936,116 +934,10 @@ stack_id=stack-id
 			},
 			wantErr: errors.E(genfile.ErrContentEval),
 		},
-		{
-			name:  "lets.map label conflicts with lets name",
-			stack: "/stack",
-			configs: []hclconfig{
-				{
-					path: "/stack/genfile.tm",
-					add: GenerateFile(
-						Labels("test"),
-						Lets(
-							Str("name", "value"),
-							Map(
-								Labels("name"),
-								Expr("for_each", "[]"),
-								Str("key", "a"),
-								Str("value", "a"),
-							),
-						),
-						Expr("content", "let.name"),
-					),
-				},
-			},
-			wantErr: errors.E(lets.ErrRedefined),
-		},
-		{
-			name:  "lets with map block",
-			stack: "/stack",
-			configs: []hclconfig{
-				{
-					path: "/stack/genfile.tm",
-					add: GenerateFile(
-						Labels("test"),
-						Lets(
-							Map(
-								Labels("var"),
-								Expr("for_each", `["a", "b", "c"]`),
-								Expr("key", "element.new"),
-								Expr("value", "element.new"),
-							),
-						),
-						Str("content", "${let.var.a}-${let.var.b}-${let.var.c}"),
-					),
-				},
-			},
-			want: []result{
-				{
-					name: "test",
-					file: genFile{
-						condition: true,
-						body:      "a-b-c",
-					},
-				},
-			},
-		},
-		{
-			name:  "lets with map using iterator",
-			stack: "/stack",
-			configs: []hclconfig{
-				{
-					path: "/stack/genfile.tm",
-					add: GenerateFile(
-						Labels("test"),
-						Lets(
-							Map(
-								Labels("var"),
-								Expr("iterator", "el"),
-								Expr("for_each", `["a", "b", "c"]`),
-								Expr("key", "el.new"),
-								Expr("value", "el.new"),
-							),
-						),
-						Str("content", "${let.var.a}-${let.var.b}-${let.var.c}"),
-					),
-				},
-			},
-			want: []result{
-				{
-					name: "test",
-					file: genFile{
-						condition: true,
-						body:      "a-b-c",
-					},
-				},
-			},
-		},
 	}
 
 	for _, tcase := range tcases {
 		testGenfile(t, tcase)
-	}
-}
-
-func TestGenFileLetsMapSchemaErrors(t *testing.T) {
-	for _, maptc := range maptest.SchemaErrorTestcases() {
-		tc := testcase{
-			name:  "genfile with lets and " + maptc.Name,
-			stack: "/stack",
-			configs: []hclconfig{
-				{
-					path: "/stack/file.tm",
-					add: GenerateFile(
-						Labels("test"),
-						Lets(
-							maptc.Block,
-						),
-					),
-				},
-			},
-			wantErr: errors.E(hcl.ErrTerramateSchema),
-		}
-		testGenfile(t, tc)
 	}
 }
